mail_merge/requirements.txt
flask==3.0.3
openpyxl==3.1.2
python-dateutil==2.9.0
werkzeug==3.0.3
requests==2.32.3
gunicorn==22.0.0
